Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float num;
-    scanf("%f",&num);
+    double num;
+    scanf("%lf",&num);
     printf("You entered: %f", num);
     return 0;
 }
